FileAttachment virus checking is handled in Asset Manager
We want to switch the storage engine from
AssetManagerAndQuarantinedFileStorage to AssetManagerStorage for the
AttachmentUploader. When we do so virus checking, which currently
happens in Whitehall and in Asset Manager, will only happen in Asset
Manager. To make the change simpler, we can remove the validation
around virus checking of attachments first, before we switch the
storage engine.

Setting FileAttachment#could_contain_viruses? to `false` reflects that
we are no longer interested in checking File attachments for viruses.

We can also remove the test for validity as it is no longer
relevant. It uses the `skip_virus_status_check` method (which is only
called from test code) to bypass the method we have just simplified.