Inline Admin::EditionsHelper#attachment_virus_status
This method always returns nil so the return value can be inlined and
the markup simplified.

Note with this change there will be no indication to the admin user
of the virus status of the attachment. However with the move to Asset
Manager for storing these assets there will be no indication (other
than the placeholder image) of the virus checking status.

A few things make me comfortable about making this change:
1) We haven't had a positive virus check for several years
2) Asset Manager won't serve assets to the public when the virus check
fails
3) Virus scanning is much quicker in Asset Manager, which uses a
single job to scan each uploaded asset, than Whitehall which relies on
a cronjob.

If we find that the Asset Manager virus checking is too slow and
confusing for admin users, we could consider adding some javascript
polling to the admin interface to query Asset Manager for the status
of the uploaded asset.