AttachmentData#virus_status is always pending
With the planned move to use AssetManagerStorage for
AttachmentUploader all virus checking will happen in Asset Manager
rather than Whitehall. To make it easier to make this change, we can
set the `virus_status` of AttachmentData to always return `pending`
which means in practice that this class is only concerned with files
that are in the carrierwave temporary directory before they are
uploaded to asset manager.