Make AttachmentsController#show infected behaviour more explicit
There are a number of scenarios in which the response is a 404 Not
Found. I'm hoping to extract the logic for each of these scenarios into
separate guard conditions to make things more explicit and to more
closely match the logic in the Asset Manager media controller download
actions. This is a first step in that direction.