Move methods used by AttachmentAccessLimitedUpdater to AttachmentData
This brings the updater more into line with the other updaters and means
we can make use of the existing AttachmentData#last_attachable method to
avoid having to have the delegating methods on Attachment.

This is also a step towards having the updater operate on an
AttachmentData rather than an Attachment which will also bring this
updater more into line with the other updaters.