Extract taxonomy functional tests into concern
Functional tests for the topic taxonomy have only been written for
publications. I've extracted these tests into a concern, which then
allowed me to apply them to all of the other document types that
support topic taxons.

The functional tests for topic taxons had some duplication: the logic
for DfE and SFA is the same. I've removed the redundant test and
simplified the remaining test to focus on the feature and not the
specific test data.